Rename package: `utils` -> `util`
