+++ screen-reports/src/main/java/com/ivy/reports/ReportViewModel.kt
@@ -9,7 +9,7 @@ import com.ivy.core.data.db.entity.TransactionType
 import com.ivy.core.data.model.Account
 import com.ivy.core.data.model.Category
 import com.ivy.core.data.model.Transaction
-import com.ivy.core.utils.stringRes
+import com.ivy.core.util.stringRes
 import com.ivy.frp.filterSuspend
 import com.ivy.frp.viewmodel.FRPViewModel
 import com.ivy.frp.viewmodel.readOnly
