+++ screen-search/src/main/java/com/ivy/search/SearchScreen.kt
@@ -24,7 +24,7 @@ import androidx.lifecycle.viewmodel.compose.viewModel
 import com.ivy.core.data.model.Account
 import com.ivy.core.data.model.Category
 import com.ivy.core.data.model.TransactionHistoryItem
-import com.ivy.core.utils.stringRes
+import com.ivy.core.util.stringRes
 import com.ivy.legacy.IvyWalletPreview
 import com.ivy.legacy.data.AppBaseData
 import com.ivy.legacy.ui.SearchInput
