+++ temp-legacy-code/src/main/java/com/ivy/legacy/IvyComposeApp.kt
@@ -13,7 +13,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalView
 import com.ivy.core.RootScreen
-import com.ivy.core.utils.appContext
+import com.ivy.core.util.appContext
 import com.ivy.design.IvyContext
 import com.ivy.design.api.IvyDesign
 import com.ivy.design.api.ivyContext
